requirements.txt
streamlit
pymupdf
google-generativeai
requests
protobuf
packaging
